Minor edits.

+++ ort_ws/src/elysium/elysium/geo_locator.py
@@ -123,14 +123,14 @@ class GeoLocator(Node):
             ),
             child_frame_id="base_link",
             pose=PoseWithCovariance(
-                pose=Pose(position=Point(x=self.x_pos, y=self.y_pos, z=self.z_pos))
+                pose=Pose(position=Point(x=float(self.x_pos), y=float(self.y_pos), z=float(self.z_pos)))
             ),
             twist=TwistWithCovariance(
                 twist=Twist(
                     linear=Vector3(
-                        x=self.dx / self.dt,
-                        y=self.dy / self.dt,
-                        z=(self.z_pos - self.z_prev_) / self.distance_sensor_dt_,
+                        x=float(self.dx / self.dt),
+                        y=float(self.dy / self.dt),
+                        z=float((self.z_pos - self.z_prev_) / self.distance_sensor_dt_)
                     )
                 )
             ),
